# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Valo API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DailyLog {
  id: number;
  userId: string;
  date: string;
  /** @nullable */
  sleepHours?: number | null;
  /** @nullable */
  hrv?: number | null;
  /** @nullable */
  restingHeartRate?: number | null;
  /** @nullable */
  steps?: number | null;
  /** @nullable */
  workoutType?: string | null;
  /** @nullable */
  workoutDuration?: number | null;
  /** @nullable */
  workoutEffort?: number | null;
}

export type DailyLogOrNull = DailyLog | null;

export interface DailyLogInput {
  /** @nullable */
  sleepHours?: number | null;
  /** @nullable */
  hrv?: number | null;
  /** @nullable */
  restingHeartRate?: number | null;
  /** @nullable */
  steps?: number | null;
  /** @nullable */
  workoutType?: string | null;
  /** @nullable */
  workoutDuration?: number | null;
  /** @nullable */
  workoutEffort?: number | null;
}

export interface MoodEntry {
  id: number;
  userId: string;
  date: string;
  score: number;
  /** @nullable */
  note?: string | null;
}

export interface MoodInput {
  score: number;
  /** @nullable */
  note?: string | null;
}

export interface Goal {
  id: number;
  userId: string;
  title: string;
  /** @nullable */
  targetDate?: string | null;
  progressPercent: number;
  category?: string;
}

export interface GoalInput {
  title: string;
  /** @nullable */
  targetDate?: string | null;
  progressPercent?: number;
  category?: string;
}

export interface GoalUpdate {
  title?: string;
  /** @nullable */
  targetDate?: string | null;
  progressPercent?: number;
  category?: string;
}

export interface Habit {
  id: number;
  userId: string;
  name: string;
  streak: number;
  completedToday: boolean;
  category?: string;
}

export interface HabitInput {
  name: string;
  category?: string;
}

export interface HabitUpdate {
  name?: string;
  completedToday?: boolean;
  streak?: number;
  category?: string;
}

export interface InsightEntry {
  id: number;
  userId: string;
  date: string;
  label: string;
  content: string;
  /** @nullable */
  followUpQuestion?: string | null;
}

export interface LogEntry {
  id: number;
  userId: string;
  date: string;
  type: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  value?: string | null;
}

export interface LogEntryInput {
  type: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  value?: string | null;
}

export interface Dashboard {
  /** @nullable */
  sleepHours?: number | null;
  /** @nullable */
  hrv?: number | null;
  /** @nullable */
  restingHeartRate?: number | null;
  /** @nullable */
  steps?: number | null;
  healthScore: number;
  workScore: number;
  relationshipScore: number;
  healthStatus: string;
  workStatus: string;
  relationshipStatus: string;
}