# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Valo API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: daily-logs
    description: Daily health metric logs
  - name: moods
    description: Mood entries
  - name: goals
    description: User goals
  - name: habits
    description: Daily habits
  - name: insights
    description: AI insight cards
  - name: log-entries
    description: Activity log entries
  - name: dashboard
    description: Dashboard data
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /daily-logs/today:
    get:
      operationId: getTodayLog
      tags: [daily-logs]
      summary: Get today's daily log for the authenticated user
      responses:
        "200":
          description: Today's daily log (may be null if not yet created)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLogOrNull"

  /daily-logs:
    post:
      operationId: upsertDailyLog
      tags: [daily-logs]
      summary: Create or update today's daily log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyLogInput"
      responses:
        "200":
          description: Updated daily log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"

  /moods:
    get:
      operationId: listMoods
      tags: [moods]
      summary: List recent mood entries (last 7 days)
      responses:
        "200":
          description: List of mood entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MoodEntry"
    post:
      operationId: createMood
      tags: [moods]
      summary: Log a mood entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoodInput"
      responses:
        "201":
          description: Created mood entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodEntry"

  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /habits:
    get:
      operationId: listHabits
      tags: [habits]
      summary: List habits
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitInput"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"

  /habits/{id}:
    patch:
      operationId: updateHabit
      tags: [habits]
      summary: Update a habit (toggle completion, edit name, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdate"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
        "404":
          description: Not found
    delete:
      operationId: deleteHabit
      tags: [habits]
      summary: Delete a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /insights:
    get:
      operationId: listInsights
      tags: [insights]
      summary: List insight cards
      responses:
        "200":
          description: List of insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InsightEntry"

  /log-entries:
    get:
      operationId: listLogEntries
      tags: [log-entries]
      summary: List today's log entries
      responses:
        "200":
          description: List of log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
    post:
      operationId: createLogEntry
      tags: [log-entries]
      summary: Create a log entry (workout, meal, mood check, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogEntryInput"
      responses:
        "201":
          description: Created log entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogEntry"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get today dashboard data (pillar scores and metrics)
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    DailyLog:
      type: object
      required: [id, userId, date]
      properties:
        id:
          type: integer
        userId:
          type: string
        date:
          type: string
        sleepHours:
          type: ["number", "null"]
        hrv:
          type: ["integer", "null"]
        restingHeartRate:
          type: ["integer", "null"]
        steps:
          type: ["integer", "null"]
        workoutType:
          type: ["string", "null"]
        workoutDuration:
          type: ["integer", "null"]
        workoutEffort:
          type: ["integer", "null"]

    DailyLogOrNull:
      oneOf:
        - $ref: "#/components/schemas/DailyLog"
        - type: "null"

    DailyLogInput:
      type: object
      properties:
        sleepHours:
          type: ["number", "null"]
        hrv:
          type: ["integer", "null"]
        restingHeartRate:
          type: ["integer", "null"]
        steps:
          type: ["integer", "null"]
        workoutType:
          type: ["string", "null"]
        workoutDuration:
          type: ["integer", "null"]
        workoutEffort:
          type: ["integer", "null"]

    MoodEntry:
      type: object
      required: [id, userId, date, score]
      properties:
        id:
          type: integer
        userId:
          type: string
        date:
          type: string
        score:
          type: integer
        note:
          type: ["string", "null"]

    MoodInput:
      type: object
      required: [score]
      properties:
        score:
          type: integer
        note:
          type: ["string", "null"]

    Goal:
      type: object
      required: [id, userId, title, progressPercent]
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        targetDate:
          type: ["string", "null"]
        progressPercent:
          type: integer
        category:
          type: string

    GoalInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        targetDate:
          type: ["string", "null"]
        progressPercent:
          type: integer
        category:
          type: string

    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        targetDate:
          type: ["string", "null"]
        progressPercent:
          type: integer
        category:
          type: string

    Habit:
      type: object
      required: [id, userId, name, streak, completedToday]
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        streak:
          type: integer
        completedToday:
          type: boolean
        category:
          type: string

    HabitInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        category:
          type: string

    HabitUpdate:
      type: object
      properties:
        name:
          type: string
        completedToday:
          type: boolean
        streak:
          type: integer
        category:
          type: string

    InsightEntry:
      type: object
      required: [id, userId, date, label, content]
      properties:
        id:
          type: integer
        userId:
          type: string
        date:
          type: string
        label:
          type: string
        content:
          type: string
        followUpQuestion:
          type: ["string", "null"]

    LogEntry:
      type: object
      required: [id, userId, date, type, title]
      properties:
        id:
          type: integer
        userId:
          type: string
        date:
          type: string
        type:
          type: string
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        value:
          type: ["string", "null"]

    LogEntryInput:
      type: object
      required: [type, title]
      properties:
        type:
          type: string
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        value:
          type: ["string", "null"]

    Dashboard:
      type: object
      required: [healthScore, workScore, relationshipScore, healthStatus, workStatus, relationshipStatus]
      properties:
        sleepHours:
          type: ["number", "null"]
        hrv:
          type: ["integer", "null"]
        restingHeartRate:
          type: ["integer", "null"]
        steps:
          type: ["integer", "null"]
        healthScore:
          type: integer
        workScore:
          type: integer
        relationshipScore:
          type: integer
        healthStatus:
          type: string
        workStatus:
          type: string
        relationshipStatus:
          type: string