# lib/api-zod/src/generated/types/dailyLogInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Valo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DailyLogInput {
  /** @nullable */
  sleepHours?: number | null;
  /** @nullable */
  hrv?: number | null;
  /** @nullable */
  restingHeartRate?: number | null;
  /** @nullable */
  steps?: number | null;
  /** @nullable */
  workoutType?: string | null;
  /** @nullable */
  workoutDuration?: number | null;
  /** @nullable */
  workoutEffort?: number | null;
}