# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Valo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Dashboard {
  /** @nullable */
  sleepHours?: number | null;
  /** @nullable */
  hrv?: number | null;
  /** @nullable */
  restingHeartRate?: number | null;
  /** @nullable */
  steps?: number | null;
  healthScore: number;
  workScore: number;
  relationshipScore: number;
  healthStatus: string;
  workStatus: string;
  relationshipStatus: string;
}