# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Valo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GoalInput {
  title: string;
  /** @nullable */
  targetDate?: string | null;
  progressPercent?: number;
  category?: string;
}