# lib/api-zod/src/generated/types/habit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Valo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Habit {
  id: number;
  userId: string;
  name: string;
  streak: number;
  completedToday: boolean;
  category?: string;
}