# lib/api-zod/src/generated/types/habitUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Valo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HabitUpdate {
  name?: string;
  completedToday?: boolean;
  streak?: number;
  category?: string;
}