# lib/api-zod/src/generated/types/insightEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Valo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InsightEntry {
  id: number;
  userId: string;
  date: string;
  label: string;
  content: string;
  /** @nullable */
  followUpQuestion?: string | null;
}