# lib/api-zod/src/generated/types/logEntryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Valo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LogEntryInput {
  type: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  value?: string | null;
}