# lib/api-zod/src/generated/types/moodEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Valo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MoodEntry {
  id: number;
  userId: string;
  date: string;
  score: number;
  /** @nullable */
  note?: string | null;
}